updated getnextline to work with zero terminated files

+++ ft_memmove.c
@@ -6,7 +6,7 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/01/11 17:35:27 by pholster       #+#    #+#                */
-/*   Updated: 2019/04/11 21:41:16 by pholster      ########   odam.nl         */
+/*   Updated: 2019/07/14 14:45:26 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_memmove(void *dst, const void *src, size_t len)
 {
-	if ((unsigned long)dst - (unsigned long)src >= len)
+	if ((size_t)dst - (size_t)src >= len)
 		ft_memcpy(dst, src, len);
 	else
 		ft_memrcpy(dst, src, len);
